+++ application/shared/controller/timeclock/payperiod.php
@@ -1,18 +1,18 @@
 <?php
 /**
- * @Author: Kenyon Haliwell
- * @Date Created: 11/13/13
- * @Date Modified: 12/11/13
- * @Purpose: Controller for pay periods
- * @Version: 2.0
+ * Author: Kenyon Haliwell
+ * Date Created: 11/13/13
+ * Date Modified: 12/18/13
+ * Purpose: Controller for pay periods
+ * Version: 2.0
  */
 
 /**
- * @Purpose: Controller for pay periods
+ * Purpose: Controller for pay periods
  */
 class timeclock_payperiod extends controller {
     /**
-     * @Purpose: Primarily used to load models based on $this->_dependencies;
+     * Purpose: Primarily used to load models based on $this->_dependencies;
      */
     public function load_dependencies($dependencies) {
         foreach ($dependencies as $dependency) {
@@ -23,7 +23,7 @@ class timeclock_payperiod extends controller {
     }
     
     /**
-     * @Purpose: This function is used to determin if the user is logged in or not
+     * Purpose: This function is used to determin if the user is logged in or not
      */
     protected function is_logged_in() {
         $this->load_dependencies(array('loggedIn'));
@@ -32,12 +32,12 @@ class timeclock_payperiod extends controller {
     }
     
     /**
-     * @Purpose: Default function to be run when class is called
+     * Purpose: Default function to be run when class is called
      */
     public function index() {}
     
     /**
-     * @Purpose: Used to punch in and out using the RESTFUL API
+     * Purpose: Used to punch in and out using the RESTFUL API
      */
     public function tx($uid) {
         $this->load_dependencies(array('payPeriod'));
@@ -66,80 +66,12 @@ class timeclock_payperiod extends controller {
     }
     
     /**
-     * @Purpose: Used to send data back using the RESTFUL API
+     * Purpose: Placeholder - what would be useful information to return about a pay period?
      */
-    public function rx($uid, $data, $pay_period='current') {
-        $this->load_dependencies(array('payPeriod', 'settings'));
-        
-        $check_employee_id = $this->sys->db->query("SELECT `employee_uid` FROM `employees` WHERE `employee_id`=:id", array(
-            ':id' => $uid
-        ));
-        
-        if (!empty($check_employee_id)) {
-            $uid = $check_employee_id[0]['employee_uid'];
-        }
-        
-        $employee = $this->sys->db->query("
-            SELECT *
-            FROM `employees` AS employee
-                JOIN `categories` AS category on category.category_id=employee.category_id
-                LEFT JOIN `job_punch` AS job on job.employee_id=employee.employee_id
-                LEFT JOIN `jobs` AS jobs on jobs.job_uid=job.job_id
-            WHERE `employee_uid`=:uid
-            ORDER BY job.punch_id DESC
-        ", array(
-            ':uid' => $uid
-        ));
-        $pay_period_query = $this->sys->db->query("SELECT `time`,`date`,`operation` FROM `employee_punch` WHERE `employee_id`=:employee_id ORDER BY `employee_punch_id` DESC", array(
-            ':employee_id' => (int) $employee[0]['employee_id']
-        ));
-        
-        $pay_period = $this->model_payPeriod->get_pay_period($pay_period);
-        $response = 'Error';
-        
-        if (empty($employee)) {
-            $this->sys->template->response = $response;
-            $this->sys->template->parse($this->sys->config->timeclock_subdirectories . '_payperiod_response');
-            
-            return False;
-        }
-
-        switch ($data) {
-            case 'name':
-                $response = $employee[0]['employee_firstname'] . ' ' . $employee[0]['employee_lastname'];
-                break;
-            case 'role':
-                $response = $employee[0]['employee_role'];
-                break;
-            case 'category':
-                $response = $employee[0]['category_name'];
-                break;
-            case 'last_job':
-                $response = $employee[0]['job_name'];
-                break;
-            case 'last_op':
-                $response = $pay_period_query[0]['operation'];
-                break;
-            case 'last_time':
-                $response = date('g:ia', $pay_period_query[0]['time']);
-                break;
-            case 'last_date':
-                $response = $pay_period_query[0]['date'];
-                break;
-            case 'total_hours':
-                $response = $this->model_payPeriod->total_hours_for_pay_period($employee[0]['employee_id'], $pay_period[0][0]);
-                break;
-            default:
-                $response = NULL;
-        }
-        
-        $this->sys->template->response = $response;
-        $this->sys->template->parse($this->sys->config->timeclock_subdirectories . '_payperiod_response');
-        return True;
-    }
+    public function rx() {}
     
     /**
-     * @Purpose: Used to create a print friendly version of a pay period
+     * Purpose: Used to create a print friendly version of a pay period
      */
     public function print_friendly($employee_id, $pay_period) {
         $this->load_dependencies(array('renderPage', 'payPeriod', 'settings'));